backend/requirements.txt
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-MySQL==1.5.2
importlib-metadata==5.0.0
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.1
PyMySQL==1.0.2
Werkzeug==2.2.2
zipp==3.8.1
timeago==1.0.16
Requests
SQLAlchemy
PyYAML==6.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.0
flask_cors
PyJWT==2.6.0
flask_bcrypt
mailjet_rest==1.3.4
flask_mail
razorpay==1.3.0
gunicorn==20.1.0
pyuwsgi==2.0.20
mysql-connector-python==8.0.31
python-dotenv==0.21.0
timeago==1.0.16
sentry-sdk[flask]==1.10.1
language_tool_python
pylatexenc
numpy
h5py==2.10.0
nltk
scikit-learn

